% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiMACC_light.R
\name{limacc}
\alias{limacc}
\title{LiMACC algorithm}
\usage{
limacc(fragment_file, contact_file, out_filename, comb_rep = "merge",
  dtype = "CHiC", peakFile4HiChIP = NULL, numG = 100,
  hcrc_pr = 0.9, fdr = 0.1, nIter = 20, maxN = 10000,
  minLen = NULL, maxLen = NULL, minDist = NULL, maxDist = 2 * 10^6,
  adj.method = "IHW", adjSeqDepth = NULL, updateMethod = "smooth")
}
\arguments{
\item{fragment_file}{the name of restriction enzyme file for CHiC/CC or binned fragment_file for HiChIP}

\item{contact_file}{contact map file(s)}

\item{out_filename}{output file name}

\item{comb_rep}{if length(contact_file)>1, merge or pool replicates (default merge for CHiC; pool for HiChIP)}

\item{dtype}{data type, CHiC/CC/HiChIP}

\item{peakFile4HiChIP}{file of coordinate of 1D peak (bed file or similar format) for HiChIP, NULL for other type of data}

\item{numG}{number of groups for limacc (default 100)}

\item{hcrc_pr}{initial guess of backgound (default 0.9)}

\item{fdr}{fdr cutoff used for the final output (default 0.05)}

\item{nIter}{maxmimum number of iterations (default 20)}

\item{maxN}{filter contacts that wiht more than maxN counts (default 100)}

\item{minLen}{filter otherEnd fragments with length smaller than minLen}

\item{maxLen}{filter otherEnd fragments with length greater than maxLen}

\item{adj.method}{method for adjusting p-value (default 'IHW')}

\item{updateMethod}{smooth or noupdate (default 'smooth')}

\item{sdepth}{the number (in million) to which the sequence depth should be normalized, default NULL means do not adjust the total
sequence depth}
}
\description{
Identify significant chromatin interations for multiple chromosome conformation capture assays
}
